Add a wire protocol version check
As per the discussion in parallaxsecond/parsec#105, it was decided that
the version check should happen in the front-end handler. Because there
is currently only one version supported (0.1), and it might stay like
this for a long time, this commit adds a simple check directly in the
request/response parser.
If multiple versions are supported, this would not suffice and a better
solution should be adopted but this is probably fine for now.

Signed-off-by: Hugues de Valon <hugues.devalon@arm.com>