Add method to check for policy and compatibility
This commit adds methods on KeyAttributes to check if a key can
perform a specific operation with a specific algorithm. Adds tests for
those methods.

This commits also implements `From` for enum types like
`NativeOperation` and `Algorithm`. It also derives serde Serialize and
Deserialize on most types as it is needed for the new Key Info Manager.

Signed-off-by: Hugues de Valon <hugues.devalon@arm.com>